feat: refactor calibration features, adding sequence, token and more detailed fragment ion matching features
