+++ pyproject.toml
@@ -32,7 +32,7 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.setuptools]
-packages = ["winnow", "winnow.scripts", "winnow.fdr", "winnow.datasets", "winnow.calibration", "winnow.compat", "winnow.configs", "winnow.configs.data_loader", "winnow.configs.fdr_method"]
+packages = ["winnow", "winnow.scripts", "winnow.fdr", "winnow.datasets", "winnow.calibration", "winnow.calibration.features", "winnow.compat", "winnow.configs", "winnow.configs.data_loader", "winnow.configs.fdr_method"]
 
 [tool.setuptools.package-data]
 winnow = ["configs/**/*.yaml", "configs/**/*.yml"]
