+++ winnow/calibration/features/constants.py
@@ -1,0 +1,4 @@
+"""Constants for calibration feature computation."""
+
+# Carbon-13 mass shift for isotopic envelope calculation
+CARBON_ISOTOPE_MASS_SHIFT = 1.00335
